Move search::FeatureValues to vespalib::FeatureValues in preparation for extending vdslib::SearchResult.
