+++ streamingvisitors/src/vespa/searchvisitor/searchvisitor.cpp
@@ -664,14 +664,14 @@ SearchVisitor::RankController::onCompletedVisiting(vsm::GetDocsumsStateCallback 
         // calculate summary features and set them on the callback object
         if (!_rankSetup->getSummaryFeatures().empty()) {
             LOG(debug, "Calculate summary features");
-            search::FeatureSet::SP sf = _rankProcessor->calculateFeatureSet();
+            vespalib::FeatureSet::SP sf = _rankProcessor->calculateFeatureSet();
             docsumsStateCallback.setSummaryFeatures(sf);
         }
 
         // calculate rank features and set them on the callback object
         if (_dumpFeatures) {
             LOG(debug, "Calculate rank features");
-            search::FeatureSet::SP rf = _dumpProcessor->calculateFeatureSet();
+            vespalib::FeatureSet::SP rf = _dumpProcessor->calculateFeatureSet();
             docsumsStateCallback.setRankFeatures(rf);
         }
     }
