+++ streamingvisitors/src/vespa/vsm/vsm/vsm-adapter.h
@@ -5,11 +5,11 @@
 #include <vespa/searchlib/query/base.h>
 #include <vespa/vsm/config/vsm-cfif.h>
 #include <vespa/config-summary.h>
-#include <vespa/searchlib/common/featureset.h>
 #include <vespa/searchsummary/docsummary/docsumwriter.h>
 #include <vespa/searchsummary/docsummary/docsumstate.h>
 #include <vespa/searchsummary/docsummary/idocsumenvironment.h>
 #include <vespa/juniper/rpinterface.h>
+#include <vespa/vespalib/util/featureset.h>
 
 using search::docsummary::ResultConfig;
 using search::docsummary::ResultClass;
@@ -28,8 +28,8 @@ class IMatchingElementsFiller;
 class GetDocsumsStateCallback : public search::docsummary::GetDocsumsStateCallback
 {
 private:
-    search::FeatureSet::SP _summaryFeatures;
-    search::FeatureSet::SP _rankFeatures;
+    vespalib::FeatureSet::SP _summaryFeatures;
+    vespalib::FeatureSet::SP _rankFeatures;
     std::unique_ptr<IMatchingElementsFiller> _matching_elements_filler;
 
 public:
@@ -37,8 +37,8 @@ public:
     void fillSummaryFeatures(GetDocsumsState& state) override;
     void fillRankFeatures(GetDocsumsState& state) override;
     std::unique_ptr<search::MatchingElements> fill_matching_elements(const search::MatchingElementsFields& fields) override;
-    void setSummaryFeatures(const search::FeatureSet::SP & sf) { _summaryFeatures = sf; }
-    void setRankFeatures(const search::FeatureSet::SP & rf) { _rankFeatures = rf; }
+    void setSummaryFeatures(const vespalib::FeatureSet::SP & sf) { _summaryFeatures = sf; }
+    void setRankFeatures(const vespalib::FeatureSet::SP & rf) { _rankFeatures = rf; }
     void set_matching_elements_filler(std::unique_ptr<IMatchingElementsFiller> matching_elements_filler);
     ~GetDocsumsStateCallback() override;
 };
