+++ vespalib/src/vespa/vespalib/util/CMakeLists.txt
@@ -31,6 +31,7 @@ vespa_add_library(vespalib_vespalib_util OBJECT
     exceptions.cpp
     execution_profiler.cpp
     executor_idle_tracking.cpp
+    featureset.cpp
     file_area_freelist.cpp
     foregroundtaskexecutor.cpp
     gate.cpp
